% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_ill_obj}
\alias{ex_ill_obj}
\title{An example illuminater oject}
\format{an illuminater object (itself a list) with 2 data frames:
\describe{
  \item{results}{An illuminate resutls by standard data frame}
  \item{roster}{a PowerSchool-style roster data frame}
}}
\usage{
ex_ill_obj
}
\description{
An illuminator object contain Illuminate results and a PowerSchool-style
roster
}
\keyword{datasets}

