% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_assessment_metadata.R
\name{il_assessment_metadata}
\alias{il_assessment_metadata}
\title{Illuminate Assessment Metadata}
\usage{
il_assessment_metadata(connection, assessment_ids)
}
\arguments{
\item{connection}{The connection object from your \code{il_connect} call}

\item{assessment_ids}{List of Illuminate IDs of the assessments you want}
}
\value{
A list of data frames for each of assessments, fields, field_responses, field_standards and field_groups
}
\description{
Get metadata for an assessment out of Illuminate.
}
\details{
The intended workflow is

- Get the metadata once, or once a week, or whatever
 - Get responses much more frequently and merge with stored metadata

There is relatively little behind-the-scenes joining,
instead returning a list of tidy data that the user can combine.
}
\examples{
c <- il_connect()
assessment_ids <- list(18672, 18657, 18677, 18679)
assess <- il_assessment_metadata(c, assessment_ids)
str(assess$assessments)
str(assess$fields)
str(assess$standards)
dbDisconnect(c)
}

