% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_assessment_responses.R
\name{il_assessment_responses}
\alias{il_assessment_responses}
\title{Illuminate Assessment Responses}
\usage{
il_assessment_responses(connection, assessment_ids)
}
\arguments{
\item{connection}{The connection object from your \code{il_connect} call}

\item{assessment_ids}{The Illuminate IDs of the assessments you want}
}
\value{
A list of data frames for each of student_assessments, student_responses and scanning_info (empty for now)
}
\description{
Get individual responses for an assessment out of Illuminate.
}
\details{
The intended workflow is

- Get the metadata once, or once a week, or whatever
 - Get responses much more frequently and merge with stored metadata

There is relatively little behind-the-scenes joining,
instead returning a list of tidy data that the user can combine.
}
\examples{
c <- il_connect()
assessment_ids <- list(18672, 18657, 18677, 18679)
resps <- il_assessment_responses(c, assessment_ids)
str(resps$students)
str(resps$student_responses)
str(resps$scanning_info)
dbDisconnect(c)

}

