% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_connect.R
\name{il_connect}
\alias{il_connect}
\title{Initiate connection to Illuminate}
\usage{
il_connect()
}
\description{
Uses your environment variables to connect to Illuminate
}
\examples{
c <- il_connect()
dbDisconnect(c)
}
\keyword{connect}
\keyword{illuminater,}

