% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_gradebook_components.R, R/il_gradebook_grades.R
\name{il_gradebook_components}
\alias{il_gradebook_assignments}
\alias{il_gradebook_categories}
\alias{il_gradebook_components}
\alias{il_gradebook_standards}
\title{Gradebook Components}
\usage{
il_gradebook_assignments(connection)

il_gradebook_categories(connection)

il_gradebook_standards(connection)
}
\arguments{
\item{connection}{The connection object from your \code{il_connect} call}
}
\value{
A dataframe with the result of the call.
}
\description{
These functions allow you to components of a gradebook out of Illuminate, without the underlying data contained within those components.
\itemize{
 \item \code{il_gradebook_assignments()}: All assignments in a gradebook
}
}

