% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_gradebook_metadata.R
\name{il_gradebook_metadata}
\alias{il_gradebook_metadata}
\title{Illuminate Gradebook Metadata}
\usage{
il_gradebook_metadata(connection)
}
\arguments{
\item{connection}{The connection object from your \code{il_connect} call}
}
\value{
A list of data frames for gradebooks, gradingperiods, and gradebooks_gradingperiods
}
\description{
Get metadata information about gradebooks, grading periods and the map between gradebooks and grading periods.
}
\examples{
c <- il_connect()
gb <- il_gradebook_metadata(c)
str(gb$gradebooks)
str(gb$gradebooks_gradingperiods)
str(gb$gradingperiods))
dbDisconnect(c)
}

