% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/illuminater_object.R
\name{illuminater}
\alias{illuminater}
\title{Title}
\usage{
illuminater(ill_con, roster, roster_table_name = NULL, ...)
}
\arguments{
\item{ill_con}{an}

\item{roster}{either a data frame with a student roster (incuding demographic)
inormation) or and `src_illuminater` object that defines a connection to
your data warehouse}

\item{roster_table_name}{if you are using a database connection then the name of
the table in the warehouse with student roster data.}

\item{...}{}
}
\value{
an illuminater object containing a roster file and illuminate assesssment
results
}
\description{
Title
}

