% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mastery_plots.R
\name{mastery_long_plot}
\alias{mastery_long_plot}
\title{Plot a grade_level mastery summary grid for a unit's worth of tests}
\usage{
mastery_long_plot(results, roster, school_id, school_name, ...)
}
\arguments{
\item{results}{illuminate results from an illuminater object}

\item{roster}{a roster to filter by}

\item{school_id}{a school id}

\item{...}{arguments passed to \code{\link[dplyr]{filter}}}

\item{school}{the name of the school with the matching id}
}
\value{
a \code{ggolot2 object}
}
\description{
Plot a grade_level mastery summary grid for a unit's worth of tests
}

