% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il_gradebook_metadata.R
\name{il_gradebook_metadata}
\alias{il_gradebook_metadata}
\title{Illuminate Gradebook Metadata}
\usage{
il_gradebook_metadata(connection)
}
\arguments{
\item{connection}{The connection object from your \code{il_connect} call}

\item{gradebook_id}{List of Illuminate IDs for each gradebook created this year}
}
\value{
A list of data frames for gradebooks, overall gradebook scores, standards and category grades, and individual assignment details
}
\description{
Get information about
1. Gradebooks
2. Overall scores
3. Category and Standard scores,
4. Assignments
}
\examples{
c <- il_connect()
gb <- il_gradebook_metadata(c)
str(gb$overall_grades)
str(gb$category_grades)
str(gb$standards_grades)
str(gb$assignments)
dbDisconnect(c)
}

